Clean up album detail screen
